No
Package name is final!
