+++ src/main/java/unrealmaed/petrock/blocks/PRBlocks.java
@@ -1,0 +1,13 @@
+package unrealmaed.petrock.blocks;
+
+import net.minecraftforge.fml.common.registry.GameRegistry;
+
+public class PRBlocks {
+
+    public static final BlockPetRock BLOCK_PET_ROCK = new BlockPetRock();
+
+    public static void load()
+    {
+        GameRegistry.registerBlock(BLOCK_PET_ROCK, BlockPetRock.getName());
+    }
+}
