+++ src/main/java/unrealmaed/petrock/client/render/model/ModelPetRockStay.java
@@ -1,0 +1,44 @@
+package unrealmaed.petrock.client.render.model;
+
+import net.minecraft.client.model.ModelBase;
+import net.minecraft.client.model.ModelRenderer;
+import net.minecraft.entity.Entity;
+
+/**
+ * Created by manmaed on 07/11/2015.
+ */
+public class ModelPetRockStay extends ModelBase {
+
+    public ModelRenderer body;
+
+    public ModelPetRockStay(int par1)
+    {
+        textureWidth = 64;
+        textureHeight = 32;
+
+        body = new ModelRenderer(this, 0, 0);
+        body.addBox(-4F, -8F, -4F, 8, 8, 8);
+        body.setRotationPoint(0F, 24F, 0F);
+        body.setTextureSize(64, 32);
+        body.mirror = true;
+        setRotation(body, 0F, 0F, 0F);
+    }
+
+    public void render(Entity entity, float f, float f1, float f2, float f3, float f4, float f5)
+    {
+        super.render(entity, f, f1, f2, f3, f4, f5);
+        setRotationAngles(f, f1, f2, f3, f4, f5, entity);
+        body.render(f5);
+    }
+
+    private void setRotation(ModelRenderer model, float x, float y, float z)
+    {
+        model.rotateAngleX = x;
+        model.rotateAngleY = y;
+        model.rotateAngleZ = z;
+    }
+
+    public void setRotationAngles(float f, float f1, float f2, float f3, float f4, float f5, Entity entity) {
+        super.setRotationAngles(f, f1, f2, f3, f4, f5, entity);
+    }
+}
