+++ src/main/java/unrealmaed/petrock/entity/EntityPetRockFollowing.java
@@ -1,0 +1,45 @@
+package unrealmaed.petrock.entity;
+
+import net.minecraft.entity.EntityAgeable;
+import net.minecraft.entity.SharedMonsterAttributes;
+import net.minecraft.entity.ai.*;
+import net.minecraft.entity.passive.EntityAnimal;
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.init.Items;
+import net.minecraft.world.World;
+
+public class EntityPetRockFollowing extends EntityAnimal {
+
+
+    public EntityPetRockFollowing(World par1World) {
+        super(par1World);
+        this.setSize(0.5F, 0.5F);
+        this.getNavigator();
+        this.tasks.addTask(0, new EntityAISwimming(this));
+        //this.tasks.addTask(0, new );
+        this.tasks.addTask(1, new EntityAIPanic(this, 2.0D));
+        this.tasks.addTask(2, new EntityAIMate(this, 1.0D));
+        this.tasks.addTask(3, new EntityAITempt(this, 1.25D, Items.wheat_seeds, false));
+        this.tasks.addTask(4, new EntityAIFollowParent(this, 1.25D));
+        this.tasks.addTask(5, new EntityAIWander(this, 1.0D));
+        this.tasks.addTask(6, new EntityAIWatchClosest(this, EntityPlayer.class, 6.0F));
+        this.tasks.addTask(7, new EntityAILookIdle(this));
+    }
+
+    protected void applyEntityAttributes()
+    {
+        super.applyEntityAttributes();
+        this.getEntityAttribute(SharedMonsterAttributes.maxHealth).setBaseValue(5.0D);
+        this.getEntityAttribute(SharedMonsterAttributes.movementSpeed).setBaseValue(0.35D);
+    }
+
+
+    public EntityPetRockFollowing spawnBabyAnimal(EntityAgeable par1EntityAgeable)
+    {
+        return new EntityPetRockFollowing(this.worldObj);
+    }
+    public EntityAgeable createChild(EntityAgeable par1EntityAgeable)
+    {
+        return this.spawnBabyAnimal(par1EntityAgeable);
+    }
+}
