+++ src/main/java/unrealmaed/petrock/proxy/CommonProxy.java
@@ -1,0 +1,20 @@
+package unrealmaed.petrock.proxy;
+
+/**
+ * Created by manmaed on 05/11/2015.
+ */
+public class CommonProxy {
+
+    public void preInit() {
+
+    }
+
+    public void init() {
+
+    }
+
+    public void renderInformation() {
+
+    }
+
+}
