--- src/main/resources/assets/petrock/blockstates/BlockPetRock.json
@@ -1,13 +1,0 @@
-{
-  "forge_marker": 1,
-  "defaults": {
-    "textures": {
-      "all": "petrock:blocks/BlockPetRock"
-    },
-    "model": "petrock:BlockPetRock"
-  },
-  "variants": {
-    "normal": [{}],
-    "inventory": [{}]
-  }
-}
