+++ src/main/resources/assets/petrock/lang/en_US.lang
@@ -1,39 +1,1 @@
-#Blocks
-tile.orePetrium.name=Petrium Ore
-tile.blockPetrium.name=Block of Petrium
-
-
-#Items
-item.matterPetrium.name=Petrium
-item.canisterLove.name=Love Canister
-item.canisterEmptyLove.name=Empty Love Canister
-item.itemPetRock.name=Pet Rock
-item.crushedStone.name=Crushed Stone
-
-
-#Keys
-key.Category=Pet Rock
-key.Hello=Hi!
-
-
-#InfoLine
-Info.wip=W.I.P
-matterPetrium.info1=This is a matter! 
-matterPetrium.info2=Press Shift For More Info.
-
-
-#Configuration
-Config.useItemPetRock=Use the Pet Rock Spawn Item. (If it is false then it will use the default spawn egg.)
-
-
-#ChatHelper
-Chat.Hello=Hi
-Chat.PetRockTime=It's Pet Rock Time!
-
-
-#Entity's
-entity.entityPetRock.name=Pet Rock
-
-
-#CreativeTab
-itemGroup.PetRock=Pet Rock
+ReBuild!
