+++ src/main/resources/mcmod.info
@@ -3,12 +3,12 @@
   "modid": "petrock",
   "name": "Pet Rock",
   "description": "A Mod about Pet Rocks",
-  "version": "0.1.0",
-  "mcversion": "1.7.10",
+  "version": "0.2.0",
+  "mcversion": "1.8.9",
   "url": "",
   "updateUrl": "",
-  "authorList": ["unrealdinnerbone"],
-  "credits": "Yunus1903, KakesRevenge, TheKayneGame, manmaed",
+  "authorList": ["Manmaed, Unrealdinnerbone"],
+  "credits": "Yunus1903, KakesRevenge, TheKayneGame",
   "logoFile": "",
   "screenshots": [],
   "dependencies": []
